requirements.txt
streamlit>=1.33
openai>=1.30
docxtpl>=0.16
python-docx>=1.1.0
pydantic>=2.7
requests>=2.32
python-dotenv>=1.0
